web/requirements.txt
bidict==0.22.0
click==8.1.3
Flask==2.2.2
Flask-SocketIO==5.3.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
Pypubsub==4.0.3
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.10.0
